a handful of package and publish chores (#312)
* chore(publish): make npm publish only run if build succeeded

* chore(package): support many versions of Angular RC

Fixes #272

* chore(package): make firebase a dependency instead of peerDep in published package

Fixes #311

* chore(package): write es modules to esm instead of es6 directory

BREAKING CHANGE:

Previously, es modules were published to the es6/ directory inside the npm
package. This change publishes them to the esm directory to be consistent
with other angular packages. Currently, the es6 directory is still published
in the npm package, but will be removed in a future release.

Fixes #310

* test: remove deprecated @angular/core/testing imports

* fix: inline sources in source maps included with npm distribution

Fixes #293

* docs(setup): add note about installing promise typings to guides

Closes #246